pyproject.toml
[project]
name = "chatad"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "firecrawl-py>=4.4.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "mcp>=1.14.0",
    "pypdf2>=3.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]
